+++ lib/torchrec.rb
@@ -9,6 +9,7 @@ require "torchrec/models/dlrm/dense_arch"
 # modules
 require "torchrec/modules/activation/swish_layer_norm"
 require "torchrec/modules/cross_net/cross_net"
+require "torchrec/modules/deepfm/deepfm"
 require "torchrec/modules/mlp/mlp"
 require "torchrec/modules/mlp/perceptron"
 require "torchrec/modules/utils"